API Remove user specifications for plot.py
These belong in their own script which calls the methods in this
module.  As was the case with main.py, having the user specifications
within the package was making version control really annoying, because
every change to those scripts for science purposes was showing up in git
as a diff.  Now only changes to the API are caught, as it should be.